+++ src/bgw_policy/policy.c
@@ -4,11 +4,13 @@
  * LICENSE-APACHE for a copy of the license.
  */
 
+#include "bgw_policy/compress_chunks.h"
 #include <postgres.h>
 
 #include "policy.h"
 #include "bgw_policy/reorder.h"
 #include "bgw_policy/drop_chunks.h"
+#include "bgw_policy/compress_chunks.h"
 #include "bgw/job.h"
 
 void
@@ -28,6 +30,11 @@ ts_bgw_policy_delete_by_hypertable_id(int32 hypertable_id)
 
 	if (policy)
 		ts_bgw_job_delete_by_id(((BgwPolicyDropChunks *) policy)->job_id);
+
+	policy = ts_bgw_policy_compress_chunks_find_by_hypertable(hypertable_id);
+
+	if (policy)
+		ts_bgw_job_delete_by_id(((BgwPolicyCompressChunks *) policy)->fd.job_id);
 }
 
 /* This function does NOT cascade deletes to the bgw_job table. */