+++ tsl/test/sql/compression_ddl.sql
@@ -280,12 +280,19 @@ WHERE hypertable.table_name like 'test1' ORDER BY hypertable.id LIMIT 1 \gset
 
 --before the drop there are 2 hypertables: the compressed and uncompressed ones
 SELECT count(*) FROM _timescaledb_catalog.hypertable hypertable;
+--add policy to make sure it's dropped later
+select add_compress_chunks_policy(:'UNCOMPRESSED_HYPER_NAME', interval '1 day');
+SELECT count(*) FROM _timescaledb_config.bgw_policy_compress_chunks;
 
 DROP TABLE :UNCOMPRESSED_HYPER_NAME;
 
 --verify that there are no more hypertable remaining
 SELECT count(*) FROM _timescaledb_catalog.hypertable hypertable;
 SELECT count(*) FROM _timescaledb_catalog.hypertable_compression;
+
+--verify that the policy is gone
+SELECT count(*) FROM _timescaledb_config.bgw_policy_compress_chunks;
+
 ROLLBACK;
 
 --create a dependent object on the compressed hypertable to test cascade behaviour
@@ -343,6 +350,12 @@ WHERE hypertable.table_name like 'test1' and chunk.compressed_chunk_id IS NOT NU
 )
 AS sub;
 
+select add_compress_chunks_policy('test1', interval '1 day');
+\set ON_ERROR_STOP 0
+ALTER table test1 set (timescaledb.compress='f');
+\set ON_ERROR_STOP 1
+
+select remove_compress_chunks_policy('test1');
 ALTER table test1 set (timescaledb.compress='f');
 
 --only one hypertable left