+++ CHANGELOG.md
@@ -23,6 +23,7 @@ that a backfill can trigger. By default, all invalidations are processed.
 **Bugfixes**
 * #1591 Fix locf treat_null_as_missing option
 * #1594 Fix error in compression constraint check
+* #1607 Delete compression policy when drop hypertable
 
 **Thanks**
 * @optijon for reporting an issue with locf treat_null_as_missing option